requirements.txt
django==2.0.2
psycopg2==2.7.3.2
django-phonenumber-field==2.0.0
djangorestframework==3.7.7
django-filter==1.1.0
Pillow==5.0.0